+++ src/winhttp.rs
@@ -797,8 +797,12 @@ pub(crate) async fn execute_request(
 
             // Set proxy Basic-auth credentials if provided.
             if let Some((username, password)) = proxy_creds {
-                abi::winhttp_set_proxy_credentials(request_handle.raw(), username, password)
-                    .url_context(url)?;
+                abi::winhttp_set_proxy_credentials(
+                    request_handle.raw(),
+                    username,
+                    password.expose(),
+                )
+                .url_context(url)?;
             }
         }
         ProxyAction::Automatic => {